requirements.txt
telegram==0.0.1
vk-api==11.9.9
redis==4.4.0
python-dotenv==0.19.2
python-telegram-bot==11.1.0